changed game model to backbone
