+++ src/app.js
@@ -15,4 +15,4 @@ $(document).ready(function() {
   });
 
   gameView.render();
-})
+});
